добавляет проверку по id владельца

+++ controllers/contacts/removeById.js
@@ -1,11 +1,15 @@
 const {Contact} = require("../../model/index")
-const {NotFound, BadRequest} = require("http-errors")
+const {NotFound, BadRequest, Unauthorized} = require("http-errors")
 const mongoose = require("mongoose")
 
 const removeById = async (req, res, next) => {
   const {contactId} = req.params
   if (!mongoose.Types.ObjectId.isValid(contactId)) {
     throw new BadRequest(`Not valid id: ${contactId}`)
+  }
+  const result = await Contact.findById(contactId, "owner")
+  if (result.owner.valueOf() !== req.user._id.valueOf()) {
+    throw new Unauthorized(`Access denied`)
   }
   const data = await Contact.findByIdAndDelete(contactId)
   if (!data) {
